+++ hyperfit/src/main/java/org/hyperfit/resource/HyperResource.java
@@ -8,7 +8,7 @@ import java.util.LinkedHashSet;
 
 /**
  * Represents a Hypermedia Resource.  Provides a set of functions for working with Hypermedia Resources
- * namely the retrieval of data and hypermedia controls like links and forms.
+ * namely the retrieval of data from fields and working with hypermedia controls like links and forms.
  *
  * It is expected that different hypermedia formats have an implementation of this interface to add support
  * for that format within hyperfit.  Implementors can make use of the BaseHyperResource abstract class to
@@ -51,24 +51,29 @@ public interface HyperResource {
      * Traverses the resource with the given path and returns the destination as the given class
      * @deprecated replaced with the getAsPath that takes a boolean indicating if a null should be returned
      * @param classToReturn the class to attempt to convert the destination into
-     * @param path the path with in the resource to locate the destination.  This is resource type specific, but usually is . notation
+     * @param dataFieldPath the path with in the resource to locate the destination.  This is resource type specific, but usually is . notation
      * @param <T> The type to return
      * @return the class if the path was valid, found and convertible.  Otherwise throws exceptions
      */
     @Deprecated
-    <T> T getPathAs(Class<T> classToReturn, String... path);
+    <T> T getPathAs(Class<T> classToReturn, String... dataFieldPath);
 
     /**
      * Traverses the resource with the given path and returns the destination as the given class
      * @param classToReturn the class to attempt to convert the destination into
      * @param nullWhenMissing when true a path leading to no destination returns null, otherwise a path with no destination caues an exception to be thrown
-     * @param path the path with in the resource to locate the destination.  This is resource type specific, but usually is . notation
+     * @param dataFieldPath the path with in the resource to locate the destination.  This is resource type specific, but usually is . notation
      * @param <T> The type to return
      * @return the class if the path was valid, found and convertible.  null if the path was not found and nullWhenMissing is true.  Otherwise throws exceptions
      */
-    <T> T getPathAs(Class<T> classToReturn, boolean nullWhenMissing, String... path);
+    <T> T getPathAs(Class<T> classToReturn, boolean nullWhenMissing, String... dataFieldPath);
 
     boolean hasPath(String... path);
+
+    /**
+     * @return an array containing all the names of the data fields on this resource
+     */
+    String[] getDataFieldNames();
 
     boolean canResolveLinkLocal(String relationship);